requirement.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
dash==2.4.1
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
Deprecated==1.2.13
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.8.1
django-paypal==2.0
django-plotly-dash==1.7.2
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
dpd-components==0.1.0
Flask==2.1.2
Flask-Compress==1.12
future==0.18.2
gcloud==0.18.3
googleapis-common-protos==1.56.4
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.3
numpy==1.22.4
oauth2client==4.1.3
pandas==1.4.2
Pillow==9.1.1
plotly==5.8.0
prompt-toolkit==3.0.29
protobuf==4.21.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.3
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
sqlparse==0.3.1
tenacity==8.0.1
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.8.0
zope.interface==5.4.0
